sendTemplate api

+++ controller/MailManager/index.js
@@ -5,42 +5,26 @@ const transporter = nodemailer.createTransport({
     service: 'gmail',
     auth: {
         user: 'lionelthegoatmessi@gmail.com',
-        pass: 'lionelthegoatmessi10'
+        pass: 'lionelmessitg10'
     }
 });
 
-const sendMail = (details, attachments = []) => {
-    // console.log(attachments)
-    if(attachments.length) {
-        mailOptions = {
-            from: details.from,
-            to: details.to,
-            subject: 'Requested Data',
-            text: 'Requested Data',
-            attachments: attachments
-        }
-    } else {
-        mailOptions = {
-            from: details.from,
-            to: details.to,
-            subject: 'Requested Data',
-            text: 'Requested Data',
-        }
-    }
-
+const sendMail = (mailOptions, flag) => {
     transporter.sendMail(mailOptions, function(error, info) {
         if (error) {
             console.log(error);
         } else {
             console.log('Email sent: ' + info.response);
-            fs.unlink('./data.csv', (err) => {
-                if (err) {
-                  console.error(err)
-                  return
-                } else {
-                    console.log('.csv file deleted')
-                }
-            })
+            if(flag) {
+                fs.unlink('./data.csv', (err) => {
+                    if (err) {
+                    console.error(err)
+                    return
+                    } else {
+                        console.log('.csv file deleted')
+                    }
+                })
+            }
         }
     });
 }
